JBC2 and more refactoring
